// src/common/enums/error-code.enum.ts
export enum ErrorCode {
  VALIDATION_ERROR = 'VALIDATION_ERROR',
  COURIER_NOT_FOUND = 'COURIER_NOT_FOUND',
  COURIER_API_ERROR = 'COURIER_API_ERROR',
  COURIER_TIMEOUT = 'COURIER_TIMEOUT',
  COURIER_AUTH_FAILED = 'COURIER_AUTH_FAILED',
  ORDER_NOT_FOUND = 'ORDER_NOT_FOUND',
  ORDER_ALREADY_EXISTS = 'ORDER_ALREADY_EXISTS',
  ORDER_CANNOT_CANCEL = 'ORDER_CANNOT_CANCEL',
  BULK_PARTIAL_FAILURE = 'BULK_PARTIAL_FAILURE',
  INTERNAL_ERROR = 'INTERNAL_ERROR',
  RATE_LIMIT_EXCEEDED = 'RATE_LIMIT_EXCEEDED',
}